fixed MSVC compile warnings

+++ tutorials/dynamic_scene/dynamic_scene_device.cpp
@@ -281,7 +281,7 @@ void animateSphere (int id, float time)
 #if 1 // enables parallel execution
   parallel_for(size_t(0),size_t(numPhi+1),[&](const range<size_t>& range) {
     for (size_t i=range.begin(); i<range.end(); i++)
-      animateSphere(i,vertices,rcpNumTheta,rcpNumPhi,pos,r,f);
+      animateSphere((int)i,vertices,rcpNumTheta,rcpNumPhi,pos,r,f);
   }); 
 #else
   for (unsigned int phi=0; phi<numPhi+1; phi++) for (int theta=0; theta<numTheta; theta++)
@@ -320,7 +320,7 @@ extern "C" void device_render (int* pixels,
   const int numTilesY = (height+TILE_SIZE_Y-1)/TILE_SIZE_Y;
   parallel_for(size_t(0),size_t(numTilesX*numTilesY),[&](const range<size_t>& range) {
     for (size_t i=range.begin(); i<range.end(); i++)
-      renderTileTask(i,pixels,width,height,time,camera,numTilesX,numTilesY);
+      renderTileTask((int)i,pixels,width,height,time,camera,numTilesX,numTilesY);
   }); 
 }
